Bitmain Ant Miner D3 X11 Dash crypto Miner 19.3GH/s ASIC w/
Bitmain antminer up for RENT. Brand new in the box plus A+++ battery. Can be available for setup if needed. It can be converted straight to Bitcoin or can only mine as Dash. So it's a dual processor. Bitcoin is the future of money and alot of people are investing in it.. why not be your own bank and chose how your money can be control without the permission of any Banks.
Brand new antminer. Never been used! With batteries. All brand new. Its the D3 X11 Dash Crypto Miner 19.3GH/s ASIC w/APW... can be delivered if the right offer comes in. It can also mine Bitcoin or be converted from Dash to Bitcoin depending on the market or on what you intend doing. First come First serve as its coming with a Bargain.
https://expertminingoutliers.com/
Email- sales@expertminingoutliers.com
Contact; +380 95 617 0523